requirements.txt
datasets>=3.0.0
pandas>=2.0.0
pyarrow>=14.0.0
pytest>=8.0.0
openai>=1.40.0
python-dotenv>=1.0.0
httpx>=0.27.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
